shifted the function & interface to dateUtils.ts
